split Writable → Creatable, Deletable, Updatable add Create, Delete, Update traits fix current implementations f̶i̶x̶ ̶t̶e̶s̶t̶s̶ wow tests still pass
