+++ src/Resource/ApiToken/Endpoint.php
@@ -11,13 +11,22 @@ namespace Nexcess\Sdk\Resource\ApiToken;
 
 use Nexcess\Sdk\ {
   Resource\ApiToken\Resource,
-  Resource\WritableEndpoint
+  Resource\Create,
+  Resource\Createable,
+  Resource\Delete,
+  Resource\Deleteable,
+  Resource\Endpoint as BaseEndpoint,
+  Resource\Update,
+  Resource\Updateable
 };
 
 /**
  * API actions for portal Login.
  */
-class Endpoint extends WritableEndpoint {
+class Endpoint
+  extends BaseEndpoint
+  implements Creatable, Deletable, Updatable {
+  use Create, Delete, Update;
 
   /** {@inheritDoc} */
   protected const _URI = 'api-token';
