+++ src/Resource/Creatable.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * @package Nexcess-SDK
+ * @license https://opensource.org/licenses/MIT
+ * @copyright 2018 Nexcess.net, LLC
+ */
+
+declare(strict_types  = 1);
+
+namespace Nexcess\Sdk\Resource;
+
+use Nexcess\Sdk\ {
+  Resource\Modelable as Model,
+  Resource\Readable
+};
+
+/**
+ * Interface for API endpoints which can create new resources.
+ */
+interface Creatable extends Readable {
+
+  /**
+   * Creates a new item.
+   *
+   * @param array $data Map of values for new item
+   * @return Model
+   * @throws ApiException If request fails
+   */
+  public function create(array $data) : Model;
+}
